requirements.txt
flask
Pillow==5
flask-bootstrap
gevent
gunicorn[eventlet]